requirements.txt
fastapi
pydantic
sqlalchemy
pymysql
uvicorn
tkinter
pillow
opencv-python
mysql-connector-python